# perf: use size hint when creating Chunk::Collect