--- a/src/chunk.rs
+++ b/src/chunk.rs
@@ -349,11 +349,9 @@ impl<A> FromIterator<A> for Chunk<A> {
     /// assert_eq!(chunk.as_vec(), vec![1, 2, 3]);
     /// ```
     fn from_iter<T: IntoIterator<Item = A>>(iter: T) -> Self {
-        let mut chunk = Vec::default();
-        for item in iter {
-            chunk.push(item);
-        }
-        Chunk::Collect(Rc::new(RefCell::new(chunk)))
+        let vec: Vec<_> = iter.into_iter().collect();
+
+        Chunk::Collect(Rc::new(RefCell::new(vec)))
     }
 }