Improves performance of creating Chunk::Collect

before:
```
from_iter/Chunk         time:   [7.5359 µs 7.5740 µs 7.6140 µs]
                        change: [-0.3708% +1.1167% +2.5443%] (p = 0.13 > 0.05)
                        No change in performance detected.
```
after:

```
from_iter/Chunk         time:   [1.3339 µs 1.3374 µs 1.3419 µs]
                        change: [-82.399% -82.296% -82.209%] (p = 0.00 < 0.05)
                        Performance has improved.
```